moved utils out of go folder and into utils folder
